move `Middleware` into own module
